requirements.txt
aiohttp==3.7.4
beautifulsoup4==4.9.3
html5lib==1.1
discord.py==1.7.2
